extract sharedlib
