+++ core/helper/recref_handler.py
@@ -20,11 +20,11 @@ from django.template.loader import render_to_string
 
 from core import constant as core_constant
 from core.forms import ImageForm, UploadImageForm, ResourceForm, RecrefForm
-from core.helper import recref_utils, model_utils, iter_utils
+from core.helper import recref_utils, model_utils, media_service
+from sharedlib import iter_utils
 from core.helper.common_recref_adapter import RecrefFormAdapter
 from core.helper.form_utils import create_formset
 from core.models import Recref, CofkUnionImage, CofkUnionRoleCategory, CofkUnionSubject
-from core.services import media_service
 
 log = logging.getLogger(__name__)
