+++ core/helper/view_components.py
@@ -2,7 +2,7 @@ import csv
 from pathlib import Path
 from typing import Iterable, Any, Callable
 
-from core.helper import data_utils
+from sharedlib import data_utils
 
 
 class HeaderValues:
