+++ sharedlib/__init__.py
@@ -1,0 +1,4 @@
+"""
+contain module or package that have no business logic
+and can be shared across the project
+"""
