+++ sharedlib/inspect_utils.py
