+++ sharedlib/iter_utils.py
