+++ sharedlib/str_utils.py
